+++ builtins/table.h
@@ -11,7 +11,7 @@
 table_t Table_from_entries(array_t entries, const Type *type);
 void *Table_get(const table_t *t, const void *key, const Type *type);
 void *Table_get_raw(const table_t *t, const void *key, const Type *type);
-void *Table_entry(const table_t *t, uint32_t n);
+void *Table_entry(const table_t *t, int64_t n);
 void *Table_reserve(table_t *t, const void *key, const void *value, const Type *type);
 void Table_set(table_t *t, const void *key, const void *value, const Type *type);
 void Table_remove(table_t *t, const void *key, const Type *type);
@@ -22,7 +22,7 @@ bool Table_equal(const table_t *x, const table_t *y, const Type *type);
 uint32_t Table_hash(const table_t *t, const Type *type);
 CORD Table_cord(const table_t *t, bool colorize, const Type *type);
 
-void *Table_str_entry(const table_t *t, uint32_t n);
+void *Table_str_entry(const table_t *t, int64_t n);
 void *Table_str_get(const table_t *t, const char *key);
 void *Table_str_get_raw(const table_t *t, const char *key);
 void Table_str_set(table_t *t, const char *key, const void *value);
