+++ sss.c
@@ -248,7 +248,7 @@ int run_repl(gcc_jit_context *ctx)
             block = NULL;
         }
 
-        for (uint32_t i = 1; i <= Table_length(&fresh_env->global->ast_functions); i++) {
+        for (int64_t i = 1; i <= Table_length(&fresh_env->global->ast_functions); i++) {
             struct { const char *key; func_context_t *value; } *entry = Table_str_entry(&fresh_env->global->ast_functions, i);
             const char *func_addr = heap_strf("%p", entry->value->func);
             if (Table_str_get(&compiled_functions, func_addr))
@@ -275,7 +275,7 @@ int run_repl(gcc_jit_context *ctx)
 
         // Copy out the variables to GC memory
         for (table_t *bindings = fresh_env->bindings; bindings; bindings = bindings->fallback) {
-            for (uint32_t i = 1; i <= Table_length(bindings); i++) {
+            for (int64_t i = 1; i <= Table_length(bindings); i++) {
                 struct {const char *key; binding_t *value;} *entry = Table_str_entry(bindings, i);
                 binding_t *b = entry->value;
                 if (!b->sym_name || Table_str_get(env->bindings, entry->key) == b)
